+++ Editor/AnimationClipFactory.cs
@@ -1,20 +1,30 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace Humr.Editor
 {
     public static class AnimationClipFactory
     {
-        public static AnimationClip PopulateAnimationClip(RecordingTake take, Animator animator)
+        public static AnimationClip PopulateBoneRotationsClip(RecordingTake take, Animator animator)
         {
             var frameCount = take.Frames.Count;
             var totalCurves = 3 + HumanTrait.BoneName.Length * 4;
 
             var keyframes = InitializeKeyframeArrays(totalCurves, frameCount);
+            for (var frameIdx = 0; frameIdx < frameCount; frameIdx++) 
+                ProcessBoneRotationsKeyframes((BoneRotationsFrame)take.Frames[frameIdx], keyframes, frameIdx, animator);
 
-            for (var frameIdx = 0; frameIdx < frameCount; frameIdx++)
-                ProcessFrameKeyframes(take.Frames[frameIdx], keyframes, frameIdx, animator);
+            return CreateAndBindBoneRotationCurves(keyframes, animator);
+        }
 
-            return CreateAndBindCurves(keyframes, animator);
+        public static AnimationClip PopulateObjectClip(RecordingTake take, Transform transform)
+        {
+            var frameCount = take.Frames.Count;
+            var keyframes = InitializeKeyframeArrays(10, frameCount);
+            for (var frameIdx = 0; frameIdx < frameCount; frameIdx++) 
+                ProcessObjectKeyframes((ObjectFrame)take.Frames[frameIdx], keyframes, frameIdx);
+            
+            return CreateAndBindObjectCurves(keyframes, transform);
         }
 
         private static Keyframe[][] InitializeKeyframeArrays(int totalCurves, int frameCount)
@@ -24,8 +34,22 @@ namespace Humr.Editor
             return keyframes;
         }
 
-        private static void ProcessFrameKeyframes(
-            RecordingFrame frame, Keyframe[][] keyframes, int frameIdx, Animator animator)
+        private static void ProcessObjectKeyframes(ObjectFrame frame, Keyframe[][] keyframes, int frameIdx)
+        {
+            keyframes[0][frameIdx] = new Keyframe(frame.RecordTime, frame.Position.x);
+            keyframes[1][frameIdx] = new Keyframe(frame.RecordTime, frame.Position.y);
+            keyframes[2][frameIdx] = new Keyframe(frame.RecordTime, frame.Position.z);
+            keyframes[3][frameIdx] = new Keyframe(frame.RecordTime, frame.Rotation.x);
+            keyframes[4][frameIdx] = new Keyframe(frame.RecordTime, frame.Rotation.y);
+            keyframes[5][frameIdx] = new Keyframe(frame.RecordTime, frame.Rotation.z);
+            keyframes[6][frameIdx] = new Keyframe(frame.RecordTime, frame.Rotation.w);
+            keyframes[7][frameIdx] = new Keyframe(frame.RecordTime, frame.LocalScale.x);
+            keyframes[8][frameIdx] = new Keyframe(frame.RecordTime, frame.LocalScale.y);
+            keyframes[9][frameIdx] = new Keyframe(frame.RecordTime, frame.LocalScale.z);
+        }
+
+        private static void ProcessBoneRotationsKeyframes(
+            BoneRotationsFrame frame, Keyframe[][] keyframes, int frameIdx, Animator animator)
         {
             var localHipPos = ProcessHipPosition(frame.HipPosition, animator);
             keyframes[0][frameIdx] = new Keyframe(frame.RecordTime, localHipPos.x);
@@ -45,7 +69,7 @@ namespace Humr.Editor
             return armatureParent.InverseTransformPoint(rawHipPos);
         }
 
-        private static void ApplyWorldRotationsToAvatar(RecordingFrame frame, Animator animator)
+        private static void ApplyWorldRotationsToAvatar(BoneRotationsFrame frame, Animator animator)
         {
             for (var k = 0; k < HumanTrait.BoneName.Length; k++)
             {
@@ -59,7 +83,7 @@ namespace Humr.Editor
         }
 
         private static void RecordLocalRotationsToKeyframes(
-            Keyframe[][] keyframes, int frameIdx, RecordingFrame frame, Animator animator)
+            Keyframe[][] keyframes, int frameIdx, BoneRotationsFrame frame, Animator animator)
         {
             for (var k = 0; k < HumanTrait.BoneName.Length; k++)
             {
@@ -76,7 +100,7 @@ namespace Humr.Editor
             }
         }
 
-        private static AnimationClip CreateAndBindCurves(Keyframe[][] keyframes, Animator animator)
+        private static AnimationClip CreateAndBindBoneRotationCurves(Keyframe[][] keyframes, Animator animator)
         {
             var clip = new AnimationClip();
             var hipPath = PathUtils.GetHierarchyPath(animator.GetBoneTransform(0));
@@ -103,6 +127,37 @@ namespace Humr.Editor
                     new AnimationCurve(keyframes[curveBaseIndex + 3]));
             }
 
+            clip.EnsureQuaternionContinuity();
+            return clip;
+        }
+
+        private static AnimationClip CreateAndBindObjectCurves(Keyframe[][] keyframes, Transform transform)
+        {
+            var clip = new AnimationClip();
+            // var transformPath = PathUtils.GetHierarchyPath(transform);
+            var transformPath = "";
+
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localPosition.x", new AnimationCurve(keyframes[0]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localPosition.y", new AnimationCurve(keyframes[1]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localPosition.z", new AnimationCurve(keyframes[2]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localRotation.x", new AnimationCurve(keyframes[3]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localRotation.y", new AnimationCurve(keyframes[4]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localRotation.z", new AnimationCurve(keyframes[5]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localRotation.w", new AnimationCurve(keyframes[6]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localScale.x", new AnimationCurve(keyframes[7]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localScale.y", new AnimationCurve(keyframes[8]));
+            clip.SetCurve(transformPath, typeof(Transform), 
+                "localScale.z", new AnimationCurve(keyframes[9]));
+            
             clip.EnsureQuaternionContinuity();
             return clip;
         }